libMirage: CHD image: fix data access in multi-track images
Attempt to fix data access in multi-track images; have the parser
properly compute the track's start sector within CHD image (which
may differ from "true" sector address due to track padding).

Have the fragment apply the start sector to turn the fragment-relative
address into CHD-global target sector address, from which hunk
index is then computed.

Similarly, map the fragment-elative address into CHD-global target
sector address when computing the sector index inside the hunk.

+++ libmirage/images/image-chd/fragment.h
@@ -57,7 +57,7 @@ GType mirage_fragment_chd_get_type (void);
 gboolean mirage_fragment_chd_setup (
     MirageFragmentChd *self,
     shared_chd_file_t *chd_file_ptr,
-    guint32 start_sector,
+    guint64 start_sector,
     guint32 num_sectors,
     guint32 hunk_size,
     guint32 sector_size,
